Initially uploaded the module
Uploaded the module in the state that it existed at the creation of this repository.
Also uploaded examples file that I use for testing the functions as I make them. It's messy, but servers as a reference for myself and others who are PowerShell savvy lieu of proper documentation, which I indent to draw up once I have the time.